# internal/web/apidoc/docs/swagger.yaml
basePath: /
definitions:
  apidoc.JsonResponseError:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseException:
    properties:
      code:
        example: 0
        type: integer
      exceptionType:
        type: string
      message:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseOk:
    properties:
      data: {}
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  apidoc.JsonResponseUnauthorized:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Access token not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  model.AccountPositionChange:
    properties:
      folderId:
        type: string
      id:
        type: string
      position:
        type: integer
    type: object
  model.AccountResult:
    properties:
      balance:
        type: string
      currency:
        $ref: '#/definitions/model.CurrencyResult'
      folderId:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/model.UserResult'
      position:
        type: integer
      sharedAccess:
        items:
          $ref: '#/definitions/model.SharedAccess'
        type: array
      type:
        type: integer
    type: object
...
  model.CorrectionResult:
    properties:
      accountId:
        type: string
      accountRecipientId:
        type: string
      amount:
        type: string
      amountRecipient:
        type: string
      author:
        $ref: '#/definitions/model.UserResult'
      categoryId:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      payeeId:
        type: string
      tagId:
        type: string
      type:
        type: string
    type: object
  model.CreateAccountRequest:
    properties:
      balance:
        type: string
      currencyId:
        type: string
      folderId:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  model.CreateAccountResult:
    properties:
      item:
        $ref: '#/definitions/model.AccountResult'
      transaction:
        $ref: '#/definitions/model.CorrectionResult'
    type: object
...
  model.CurrencyResult:
    properties:
      code:
        type: string
      fractionDigits:
        type: integer
      id:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
...
  model.DeleteAccountRequest:
    properties:
      id:
        type: string
    type: object
  model.DeleteAccountResult:
    type: object
...
  model.GetAccountListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.AccountResult'
        type: array
    type: object
...
  model.OrderAccountListRequest:
    properties:
      changes:
        items:
          $ref: '#/definitions/model.AccountPositionChange'
        type: array
    type: object
  model.OrderAccountListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.AccountResult'
        type: array
    type: object
...
  model.SharedAccess:
    properties:
      isAccepted:
        type: integer
      role:
        type: string
      user:
        $ref: '#/definitions/model.UserResult'
    type: object
...
  model.UpdateAccountRequest:
    properties:
      balance:
        type: string
      currencyId:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  model.UpdateAccountResult:
    properties:
      item:
        $ref: '#/definitions/model.AccountResult'
      transaction:
        $ref: '#/definitions/model.CorrectionResult'
    type: object
...
  model.UserResult:
    properties:
      avatar:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Self-hosted personal finance and budgeting API
  title: Econumo API
  version: 1.0.0
paths:
...
  /api/v1/account/create-account:
    post:
      consumes:
      - application/json
      description: Creates an account (idempotent on the request id), optionally seeding
        its balance, in the given folder.
      parameters:
      - description: Create account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.CreateAccountResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Create an account
      tags:
      - Account
...
  /api/v1/account/delete-account:
    post:
      consumes:
      - application/json
      description: Soft-deletes an account. Requires ownership.
      parameters:
      - description: Delete account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeleteAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.DeleteAccountResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Delete an account
      tags:
      - Account
  /api/v1/account/get-account-list:
    get:
      description: Returns all the user's available accounts (each with owner, currency,
        folder, position, balance) in reverse order.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.GetAccountListResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Get the account list
      tags:
      - Account
...
  /api/v1/account/order-account-list:
    post:
      consumes:
      - application/json
      description: Repositions accounts and moves them between folders, returning
        the full list.
      parameters:
      - description: Order account list request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.OrderAccountListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.OrderAccountListResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Reorder the account list
      tags:
      - Account
...
  /api/v1/account/update-account:
    post:
      consumes:
      - application/json
      description: Updates an account's name/icon/currency and reconciles its balance
        via a correction transaction. Requires ownership.
      parameters:
      - description: Update account request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateAccountResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Update an account
      tags:
      - Account
...
securityDefinitions:
  Bearer:
    description: 'Opaque access token. Format: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"